Apply Hotfix KB2842230 on Windows 8 & Server 2012 templates.
There is a known issue with WinRM ignoring a customized
MaxMemoryPerShellMB on certain default installations of Windows. In this
project, the issue affects the following templates:

* win2012-datacenter-cygwin
* win2012-datacenter
* win2012-standard-cygwin
* win2012-standard
* win8x64-enterprise-cygwin
* win8x64-enterprise
* win8x64-pro-cygwin
* win8x64-pro

This commit adds a floppy-mountable script (hotfix-KB2842230.bat) which
applies the hotfix to the system. The implementation started as a port
from joefitzgerald/packer-windows but has been heavily customized to
download, extract, and apply the hotfix only to the target Windows
versions. The Microsoft-provided hotfix is a self-extracting .exe file
which has no CMD flags and thus couldn't be automated. Instead, this
implementation uses a Chocolatey package (which is a Zip format)
containing the .msu program.

For more details on KB2842230, interested parties can read:

https://support.microsoft.com/en-us/kb/2842230?wa=wsignin1.0