Further slim (#2)
* Use certificates from VBoxGuestAdditions.iso rather than oracle-cert.cer

* Remove windows updates files as part of the cleanup

Signed-off-by: Tim Smith <tsmith@chef.io>

* Use the dism command for cleaning up windows update files

This is from mwrock's blog post here http://www.hurryupandwait.io/blog/in-search-of-a-light-weight-windows-vagrant-box

Signed-off-by: Tim Smith <tsmith@chef.io>

* Use cleanmgr and remove windows update files on Windows 7

dism doesn't include the ability to cleanup components on Windows 7. Detect when we're on Windows 7 (2k8r2 as well) and do this the old school way

Signed-off-by: Tim Smith <tsmith@chef.io>

* Disable Hibernation in the powerconfig script

Signed-off-by: Tim Smith <tsmith@chef.io>

* Increase CPUs and Memory on all systems

This is still rather low, but greatly increases the speed of the builds

Signed-off-by: Tim Smith <tsmith@chef.io>

* Update UltraDefrag to 7.0.2

Signed-off-by: Tim Smith <tsmith@chef.io>

* Rename start-sshd script to start-transports

It starts all the different transports not just sshd.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Fix transport starting on Win10

I missed updating these for the new script name

Signed-off-by: Tim Smith <tsmith@chef.io>

* Fix typo

Signed-off-by: Tim Smith <tsmith@chef.io>

* Switch back to sdelete 1.61

This is 10X faster

Signed-off-by: Tim Smith <tsmith@chef.io>

* Fix minor formatting issues in the winrm logging

() didn’t show up right

Signed-off-by: Tim Smith <tsmith@chef.io>

* Disable the network discovery prompt post install on Windows 10

Signed-off-by: Tim Smith <tsmith@chef.io>

* Use the right kind of comments

I'm clearly a Linux user

Signed-off-by: Tim Smith <tsmith@chef.io>

* Sync up all the Autounattend scripts

This mostly just makes it easier to diff

Signed-off-by: Tim Smith <tsmith@chef.io>

* Always run the update/post update actions on Win 10 + uninstall 7zip

Reorder all the files and make sure the scripts are the same.

We weren't doing the post install updates everywhere
We weren't doing the updates everywhere
We were missing the 7zip uninstall in at least 1 spot
We were missing fixing the network in a few spots which makes winrm fail

Signed-off-by: Tim Smith <tsmith@chef.io>

* Always nuke the downloaded windows update files

This shaves 150 megs off the size of the windows 10 box

Signed-off-by: Tim Smith <tsmith@chef.io>

* Correctly set the vmware guest_os_type for Win10

It's Windows9...because...

Signed-off-by: Tim Smith <tsmith@chef.io>

* Update README.md

* Update sdelete URL

* Started adding Windows Server 2016 support

* Adding non-eval win2016

* Standardized VM ram as size 4096

* Upgraded chef client to 13.2.20

* Win2016 now uses 64-bit Cygwin

* Deleting the changelog for now

* Fix 32bit windows 7zip URL

Signed-off-by: Tim Smith <tsmith@chef.io>

* Fix 2 more 7zip URLs for 32bit windows

Signed-off-by: Tim Smith <tsmith@chef.io>

* Fixed iso url and checksum for win2016 eval

* Fix VBox Guest Additions install

It seems like some changes broke the path where the VirtualBox Guest
Additions were installed via a mounted CD from the Host OS. The changes
look like copy and paste errors from the downloaded ISO path. This
change fixes the paths to be correct.

* Fix AutoLogon issue with Windows 10 Enterprise build

It seems the latest release of win10 has a bug in the installer where
autologon only works once. Unfortuantely this causes packer to hang
because there are Windows Updates that require a reboot so the
post-install tasks never finish and WinRM eventually times out.

This change works around the issue by explicitly enabling autologon via
registry settings **before** update.bat has a chance to run.

* Remove EOL Windows 8

Windows 8.1 has replaced Windows 8, which went out of mainstream support Nov 2015. Since Microsoft considers 8.1 to be a SP for 8 and we can't even get eval ISOs for 8 anymore we should remove it. It's unsupported by them at this point and they're pointing everyone to at least the 8.1 upgrade.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Update the eval windows 10 to the latest release

Signed-off-by: Tim Smith <tsmith@chef.io>

* Update Chef / ChefDK to the latest versions

Signed-off-by: Tim Smith <tsmith@chef.io>

* Update vbox tools to the latest release

Signed-off-by: Tim Smith <tsmith@chef.io>

* Slim down Windows 10

Remove things we don't need for a vagrant box. Ads, personalization,
windows defender, cortana, feedback prompts.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Changed all memory and cpu definitions to use the `memory` and `cpus` user-variables defined within each json template. Also added a default-empty `guest_additions_url` variable for allowing the user to specify the guest additions iso for VirtualBox builders.

* Update Windows 10 ISOs to the 1709 Fall Creators update

Signed-off-by: Irving Popovetsky <irving@chef.io>

* Further slim down installations

Remove extra Windows 10 apps
Remove any restore points

Signed-off-by: Tim Smith <tsmith@chef.io>