Replace http links with https versions or updated versions
Every http link was checked if it was possible to find an https replacement
that works. Every https link should work, except a few things (AWS acs link?)
that do not seem to work even in the http version. Some utilities were update
to newer versions but I tried to keep the major version - to minimize the chance
of breaking builds. Not every combination of image was tested but it is already
a huge step concerning security.

The remaining http links are part of schemas or the LICENSE file and I belive
these should not be changed.