Added a upx-compressed version of curl that links with msvcrt.dll to 3rdparty/curl.exe, and updated all of the templates to include it on their floppy disk during build.
The 32-bit curl binary that's included supports the minimum number of protocols
as needed by this project, and uses SChannel for SSL. What's special about it
is that it links _only_ with msvcrt.dll so that it's supported on all versions
of Windows. It's been upx-compresesed (9) so there's more space on the floppy
which reduced it from ~550k to ~200k. The repository that this binary is built
from is at https://github.com/arizvisa/curl-windows-nodeps.